+++ src/types/index.ts
@@ -21,3 +21,4 @@ export * from './largeTextSave';
 export * from './outline';
 export * from './settingSuggestion';
 export * from './chapterEngineering';
+export * from './generationContext';